Fixing issues #130, #131 and #132.
